Move errors into a separate file
