+++ connection.go
@@ -90,15 +90,6 @@ type PrivmsgOut struct {
 	Data      string `json:"data,omitempty"`
 }
 
-type GenericError struct {
-	Description string `json:"description"`
-}
-
-type MutedError struct {
-	GenericError
-	MuteTimeLeft int64 `json:"muteTimeLeft"`
-}
-
 // Create a new connection using the specified socket and router.
 func newConnection(s *websocket.Conn, user *User, ip string) {
 	c := &Connection{
@@ -161,13 +152,7 @@ func (c *Connection) readPumpText() {
 	// Check mute status.
 	muteTimeLeft := mutes.muteTimeLeft(c)
 	if muteTimeLeft > time.Duration(0) {
-		c.EmitBlock(
-			"ERR",
-			MutedError{
-				GenericError{"muted"},
-				int64(muteTimeLeft / time.Second),
-			},
-		)
+		c.EmitBlock("ERR", NewMutedError(muteTimeLeft))
 	}
 
 	for {
@@ -408,13 +393,7 @@ func (c *Connection) canMsg(msg string, ignoresilence bool) bool {
 	if !ignoresilence {
 		muteTimeLeft := mutes.muteTimeLeft(c)
 		if muteTimeLeft > time.Duration(0) {
-			c.EmitBlock(
-				"ERR",
-				MutedError{
-					GenericError{"muted"},
-					int64(muteTimeLeft / time.Second),
-				},
-			)
+			c.EmitBlock("ERR", NewMutedError(muteTimeLeft))
 			return false
 		}
